api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.27.2"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.7"
]
